requirements.txt
Django==5.1.3
requests
python-dotenv
djangorestframework

